data_management/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
dill==0.3.8
docopt==0.6.2
filelock==3.13.1
frozenlist==1.4.1
fsspec==2023.10.0
fugashi
hojichar==0.9.0
huggingface-hub==0.20.3
idna==3.6
jsonable==0.3.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
mmh3==4.1.0
multidict==6.0.5
multiprocess==0.70.16
mwcli==0.0.3
mwtypes==0.3.2
mwxml==0.3.3
numpy==1.26.4
packaging==23.2
pandas==2.2.1
para==0.0.8
plac==1.4.3
pyarrow==15.0.0
pyarrow-hotfix==0.6
pycryptodome==3.20.0
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
requests==2.31.0
rpds-py==0.18.0
six==1.16.0
tqdm==4.66.2
typing_extensions==4.9.0
tzdata==2024.1
unidic==1.1.0
urllib3==2.2.1
wasabi==0.10.1
xxhash==3.4.1
yarl==1.9.4
